requirements.txt
astropy
numpy
PySide6
Pillow
scikit-image
matplotlib
scipy